+++ src/main/java/net/trueog/utilitiesog/Listeners/DisableEntityCrammingListener.java
@@ -1,6 +1,6 @@
 // This is free and unencumbered software released into the public domain.
 // Authors: christianniehaus, NotAlexNoyle.
-package me.barny1094875.utilitiesog.Listeners;
+package net.trueog.utilitiesog.Listeners;
 
 // Import libraries.
 import org.bukkit.event.EventHandler;
@@ -8,8 +8,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
 
-// Import required libraries.
-import me.barny1094875.utilitiesog.UtilitiesOG;
+import net.trueog.utilitiesog.UtilitiesOG;
 
 // Declare the EntityCrammingToggle Module with Bukkit Listeners.
 public class DisableEntityCrammingListener implements Listener {