+++ src/main/java/net/trueog/utilitiesog/Listeners/TogglePhantomsListener.java
@@ -1,6 +1,6 @@
 // This is free and unencumbered software released into the public domain.
 // Authors: christianniehaus, NotAlexNoyle.
-package me.barny1094875.utilitiesog.Listeners;
+package net.trueog.utilitiesog.Listeners;
 
 // Import libraries.
 import org.bukkit.Bukkit;
@@ -11,7 +11,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.CreatureSpawnEvent;
 
-import me.barny1094875.utilitiesog.UtilitiesOG;
+import net.trueog.utilitiesog.UtilitiesOG;
 
 // Declare the DisablePhantomSpawns Module with Bukkit Listeners.
 public class TogglePhantomsListener implements Listener {