+++ src/main/java/net/trueog/utilitiesog/commands/ToggleCrammingCommand.java
@@ -1,6 +1,6 @@
 // This is free and unencumbered software released into the public domain.
 // Authors: christianniehaus, NotAlexNoyle.
-package me.barny1094875.utilitiesog.commands;
+package net.trueog.utilitiesog.commands;
 
 // Import libraries.
 import org.bukkit.command.Command;
@@ -9,8 +9,8 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 
-import me.barny1094875.utilitiesog.UtilitiesOG;
-import me.barny1094875.utilitiesog.utils.TextUtils;
+import net.trueog.utilitiesog.UtilitiesOG;
+import net.trueog.utilitiesog.utils.TextUtils;
 
 // Declare the /togglecramming command in the DisableEntityCramming Module with Bukkit Commands.
 public class ToggleCrammingCommand implements CommandExecutor {
@@ -53,7 +53,7 @@ public class ToggleCrammingCommand implements CommandExecutor {
 		else {
 
 			// Send a detailed error message to the server console.
-			TextUtils.consoleUseErrorMessage(sender, cmd.getName(), toggleCrammingPermission);
+			TextUtils.logToConsole(cmd.getName() + " " + toggleCrammingPermission);
 
 		}